This report, by its very length, defends itself against the risk of being read

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -9,7 +9,7 @@
 
 int _strlen(char *str)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
